remove TUI
